fix(pairing): green — no-op re-confirm preserves original attribution
Review findings on PR 54 (issue 45), fixed here since no GitHub issue
tracks this fixup:

1. confirmPairing() (modules/pairing/state.ts) now reads the existing
   row's state alongside its supplierUrl (same .for("update") lock as
   before) and short-circuits to a no-op ("ok", no upsert, no audit write)
   exactly when the row is ALREADY 'potvrdene' AND the resulting URL is
   UNCHANGED — no new decision was made, so the original confirmedBy/
   confirmedAt must survive untouched. A genuine URL change (manual
   override to a different address) still goes through the normal
   upsert+audit path, moving attribution to the current actor, because
   that IS a new decision. Verified against the CHECK constraint
   (pairing_confirmation_ck): the no-op path performs no write at all, so
   it cannot violate it.

2. Web side (PairingSection.tsx): the confirm button is now also disabled
   once the row is already in the confirmed state — an already-confirmed
   row is corrected only via the manual-address button (which still
   writes a real URL change). This is UX only, not the real fix — the
   server-side guard above is what actually prevents the theft, since the
   API is the authority (a direct call or stale cached UI could bypass a
   UI-only guard).

3. Dead code removed: the empty-string branch of the missing-URL check in
   the same function was unreachable — the only writer of
   pairings.supplier_url is this very function, fed either null or a
   zod-validated non-empty URL string (pairing-routes.ts's request body
   schema), so an empty string can never reach this check. Reduced to the
   null case only; the missing_url outcome and its existing test are
   unaffected.

GREEN: both new integration tests in the new
pairing-reconfirm.integration.test.ts (split out from
pairing-http.integration.test.ts, which was pushing the 400-line
max-lines ESLint cap after the RED commit) now pass, plus the full
183-test integration suite and 103-test web unit suite.

New web unit test in PairingSection.test.tsx locks in the button-disable
behaviour: confirm button disabled + manual-address button enabled on an
already-confirmed row.
